+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 0.2.8
+### 0.2.8 (December 19, 2023)
 
 #### Changes
 * Changed the types in the Routing. In most cases these are still compatible because utility methods are used.